// libs/shared/src/lib/enums/env-names.enum.ts
export enum EnvNamesEnum {
  another_knowledge_base_app_port = 'ANOTHER_KNOWLEDGE_BASE_APP_PORT',
  db_host = 'DB_HOST',
  db_port = 'DB_PORT',
  db_database = 'DB_DATABASE',
  db_username = 'DB_USERNAME',
  db_password = 'DB_PASSWORD',
  db_ssl = 'DB_SSL',
  db_ssl_ca = 'DB_SSL_CA',
  redis_enabled = 'REDIS_ENABLED',
  redis_host = 'REDIS_HOST',
  redis_port = 'REDIS_PORT',
  redis_password = 'REDIS_PASSWORD',
  jwt_secret = 'JWT_SECRET',
  expires_in = 'EXPIRES_IN',
  typeorm_entities = 'TYPEORM_ENTITIES',
  typeorm_migrations = 'TYPEORM_MIGRATIONS',
  production = 'PRODUCTION',
}